move tests for building the option options list to their own file
